+++ lib/opal/nodes/rescue.rb
@@ -35,12 +35,45 @@ module Opal
 
       def compile
         push "try {"
-        line compiler.process(body_sexp, @level)
+
+        in_ensure do
+          line compiler.process(body_sexp, @level)
+        end
+
         line "} finally {"
-        line compiler.process(ensr_sexp, @level)
+
+        indent do
+          if has_rescue_else?
+            # $no_errors indicates thate there were no error raised
+            unshift "var $no_errors = true; "
+
+            # when there's a begin;rescue;else;ensure;end statement,
+            # ruby returns a result of the 'else' branch
+            # but invokes it before 'ensure'.
+            # so, here we
+            # 1. save the result of calling else to $rescue_else_result
+            # 2. call ensure
+            # 2. return $rescue_else_result
+            line "var $rescue_else_result;"
+            line "if ($no_errors) { "
+            indent do
+              line "$rescue_else_result = (function() {"
+              indent do
+                line compiler.process(compiler.returns(scope.rescue_else_sexp), @level)
+              end
+              line "})();"
+            end
+            line "}"
+            line compiler.process(ensr_sexp, @level)
+            line "if ($no_errors) { return $rescue_else_result; }"
+          else
+            line compiler.process(ensr_sexp, @level)
+          end
+        end
+
         line "}"
 
-        wrap '(function() {', '; })()' if wrap_in_closure?
+        wrap '(function() { ', '; })()' if wrap_in_closure?
       end
 
       def body_sexp
@@ -59,7 +92,7 @@ module Opal
       end
 
       def wrap_in_closure?
-        recv? or expr?
+        recv? or expr? or has_rescue_else?
       end
     end
 
@@ -69,7 +102,12 @@ module Opal
       children :body
 
       def compile
-        handled_else = false
+        scope.rescue_else_sexp = children[1..-1].detect { |sexp| sexp.type != :resbody }
+        has_rescue_handlers = false
+
+        if handle_rescue_else_manually?
+          line "var $no_errors = true;"
+        end
 
         push "try {"
         indent do
@@ -78,20 +116,42 @@ module Opal
         line "} catch ($err) {"
 
         indent do
-          children[1..-1].each_with_index do |child, idx|
-            handled_else = true unless child.type == :resbody
+          if has_rescue_else?
+            line "$no_errors = false;"
+          end
 
-            push " else " unless idx == 0
-            line process(child, @level)
+          children[1..-1].each_with_index do |child, idx|
+            # counting only rescue, ignoring rescue-else statement
+            if child.type == :resbody
+              has_rescue_handlers = true
+
+              push " else " unless idx == 0
+              line process(child, @level)
+            end
           end
 
           # if no resbodys capture our error, then rethrow
-          unless handled_else
-            push " else { throw $err; }"
-          end
+          push " else { throw $err; }"
         end
 
         line "}"
+
+        if handle_rescue_else_manually?
+          # here we must add 'finally' explicitly
+          push "finally {"
+          indent do
+            line "if ($no_errors) { "
+            indent do
+              line "return (function() {"
+              indent do
+                line compiler.process(compiler.returns(scope.rescue_else_sexp), @level)
+              end
+              line "})();"
+            end
+            line "}"
+          end
+          push "}"
+        end
 
         # Wrap a try{} catch{} into a function
         # when it's an expression
@@ -103,6 +163,13 @@ module Opal
         body_code = (body.type == :resbody ? s(:nil) : body)
         body_code = compiler.returns body_code unless stmt?
         body_code
+      end
+
+      # Returns true when there's no 'ensure' statement
+      #  wrapping current rescue.
+      #
+      def handle_rescue_else_manually?
+        !scope.in_ensure? && scope.has_rescue_else?
       end
     end